+++ Rabbit UI/src/rabbit/ui/internal/util/ProjectElement.java
@@ -42,8 +42,8 @@ public class ProjectElement extends ResourceElement {
 	}
 
 	@Override
-	public double getValue() {
-		double value = 0;
+	public long getValue() {
+		long value = 0;
 		for (ResourceElement e : getChildren()) {
 			value += e.getValue();
 		}
